Documentation for Maxvol

This Julia package provides two routines:

maxvol_generic!, generic implementation of Maxvol algorithm, which can be used for matrices of any numerical type (e.g. Rational or BigFloat).

maxvol!, LAPACK-based implementation of Maxvol algorithm, which works only for standard numerical types: Float32, Float64, ComplexF32 and ComplexF64.

Methods

Maxvol.maxvol_generic! — Function

maxvol_generic!(A, tol=1.05, maxiters=100)

Generic maxvol, that does not use any BLAS or LAPACK calls.

Can be used for special arithmetics, like BigFloat, Rational or Complex{Float16}.

Arguments:

A::Matrix: Input matrix on entry and output coefficients on exit.

tol::Float64: Stop when determinant growth is less or equal to this.

maxiters::Int: Maximum number of iterations.

Returns:

piv::Vector{Int}: Indexes of pivoted rows

niters::Int: Number of performed swap iterations

Example:

julia> using Random, LinearAlgebra, Maxvol
julia> rng = MersenneTwister(100);
julia> A = rand(rng, BigFloat, 1000, 100);
julia> C = copy(A);
julia> piv, niters = maxvol_generic!(C);
julia> norm(A-C*A[piv,:]) / norm(A)
2.030657951400512330834848952202721164346464876711701213634530270353170311161736e-76
julia> A = rand(rng, ComplexF64, 1000, 100);
julia> C = copy(A);
julia> piv, niters = maxvol_generic!(C);
julia> norm(A-C*A[piv,:]) / norm(A)
4.863490630095799e-15

See also: maxvol!

Maxvol.maxvol! — Function

maxvol!(A, tol=1.05, maxiters=100)

Uses vendor-optimized LAPACK, provided by Julia.

Supports only 4 input types: Float32 (single), Float64 (double), ComplexF32 (single complex) and ComplexF64 (double complex).

Arguments:

A::Matrix{T}: Input matrix on entry and output coefficients on exit. T must be one of Float32, Float64, ComplexF32 and ComplexF64.

tol::Float64: Stop when determinant growth is less or equal to this.

maxiters::Int: Maximum number of iterations.

Returns:

piv::Vector{Int}: Indexes of pivoted rows

niters::Int: Number of performed swap iterations

Example:

julia> using Random, LinearAlgebra, Maxvol
julia> rng = MersenneTwister(100);
julia> A = rand(rng, Float64, 1000, 100);
julia> C = copy(A);
julia> piv, niters = maxvol!(C);
julia> norm(A-C*A[piv,:]) / norm(A)
2.3975097489579994e-15
julia> A = rand(rng, ComplexF32, 1000, 100);
julia> C = copy(A);
julia> piv, niters = maxvol!(C);
julia> norm(A-C*A[piv,:]) / norm(A)
2.0852597f-6

See also: maxvol_generic!

License

This package is ditributed under BSD 3-Clause license. It can be found in the root directory of repository in LICENSE.md file.